requirements.txt
# Core Scientific Computing
numpy>=1.21.0,<2.0.0
scipy>=1.7.0,<2.0.0
pandas>=1.4.0,<2.0.0

# Machine Learning - PyTorch
torch>=1.13.0
torchvision>=0.14.0
scikit-learn>=1.0.0,<2.0.0

# Privacy-Preserving Techniques
opacus>=1.3.0

# Biomedical Signal Processing
pyedflib>=0.1.30
mne>=1.0.0
wfdb>=4.0.0

# Data & Configuration
pyyaml>=6.0
joblib>=1.1.0

# Progress & Logging
tqdm>=4.65.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0